+++ ClientApp/src/app/pages/home/home.component.html
@@ -1,11 +1,4 @@
 <main class="main">
     
-      <div id="login" class="flex flex-col gap-6 items-center">
-            
-     
-            <div id="google-btn" class="division"></div>
-          
-          
-        
-    </div>
+      
 </main>
