+++ ClientApp/src/app/pages/login/login.component.css
@@ -1,0 +1,6 @@
+.division {
+  height: 50%;
+  width: 20%;
+  align-self: center;
+  margin: 10% 10% 10% 10%;
+}
