+++ ClientApp/src/app/pages/login/login.component.html
@@ -15,21 +15,20 @@
         <label class="block text-gray-700 font-medium mb-1" for="email">Email</label>
         <div class="relative">
           <ng-icon name="heroUser" class="absolute left-3 top-3 text-gray-500"></ng-icon>
-          <input
-            type="email"
-            id="email"
-            formControlName="email"
-            class="w-full pl-10 pr-4 py-2 border rounded-lg focus:outline-none focus:ring focus:ring-green-200"
-            placeholder="Enter your email"
-            required>
+          <input type="email"
+                 id="email"
+                 formControlName="email"
+                 class="w-full pl-10 pr-4 py-2 border rounded-lg focus:outline-none focus:ring focus:ring-green-200"
+                 placeholder="Enter your email"
+                 required>
         </div>
         @if (getEmail()?.invalid && getEmail()?.touched) {
-          @if (getEmail()?.errors?.['required']) {
-            <p class="text-red-500 text-sm">Email is required.</p>
-          }
-          @if (getEmail()?.errors?.['email']) {
-            <p class="text-red-500 text-sm">Please enter a valid email.</p>
-          }
+        @if (getEmail()?.errors?.['required']) {
+        <p class="text-red-500 text-sm">Email is required.</p>
+        }
+        @if (getEmail()?.errors?.['email']) {
+        <p class="text-red-500 text-sm">Please enter a valid email.</p>
+        }
         }
       </div>
 
@@ -38,29 +37,27 @@
         <label class="block text-gray-700 font-medium mb-1" for="password">Password</label>
         <div class="relative">
           <ng-icon name="heroLockClosed" class="absolute left-3 top-3 text-gray-500"></ng-icon>
-          <input
-            type="password"
-            id="password"
-            formControlName="password"
-            class="w-full pl-10 pr-4 py-2 border rounded-lg focus:outline-none focus:ring focus:ring-green-200"
-            placeholder="Enter your password"
-            required>
+          <input type="password"
+                 id="password"
+                 formControlName="password"
+                 class="w-full pl-10 pr-4 py-2 border rounded-lg focus:outline-none focus:ring focus:ring-green-200"
+                 placeholder="Enter your password"
+                 required>
         </div>
         @if (getPassword()?.invalid && getPassword()?.touched) {
-          @if (getPassword()?.errors?.['required']) {
-            <p class="text-red-500 text-sm">Password is required.</p>
-          }
-          @if (getPassword()?.errors?.['minlength']) {
-            <p class="text-red-500 text-sm">Password must be at least 6 characters long.</p>
-          }
+        @if (getPassword()?.errors?.['required']) {
+        <p class="text-red-500 text-sm">Password is required.</p>
+        }
+        @if (getPassword()?.errors?.['minlength']) {
+        <p class="text-red-500 text-sm">Password must be at least 6 characters long.</p>
+        }
         }
       </div>
 
       <!-- Submit Button -->
-      <button
-        type="submit"
-        [disabled]="loginForm.invalid || disableSubmit"
-        class="w-full bg-green-600 text-white py-2 rounded-lg hover:bg-green-700 transition disabled:bg-gray-400">
+      <button type="submit"
+              [disabled]="loginForm.invalid || disableSubmit"
+              class="w-full bg-green-600 text-white py-2 rounded-lg hover:bg-green-700 transition disabled:bg-gray-400">
         Login
       </button>
     </form>
@@ -69,7 +66,16 @@
     <div class="mt-4 text-center">
       <a routerLink="/forgot-password" class="text-sm text-green-500 hover:underline">Forgot your password?</a>
     </div>
+
+    <hr />
+    <h3 class=" text-center" style="margin-top:10%">Continue with</h3>
+    
+    <div style="align-content:center center">
+      <div id="google-btn"></div>
+    </div>
   </div>
+
+  
 
   <!-- Footer Links -->
   <div class="mt-6 text-center">
